API/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
psycopg[binary]
python-dotenv
pydantic-settings
passlib[bcrypt]
python-jose[cryptography]
email-validator
python-multipart